Refactor out sprint timeUtils
These started as the entirety of generate.js, and now live in their own package for readability and isolation of concerns. Although they 'generate sprints' - really all they do is produce a map with start and end times, which may or may not be the entire definition of a sprint in the future. Likely not.

A big motivator for this change was the generator command template. I needed those util functions to be defined first, or template.call was undefined.

The time util functions also take the actual timestamp indicating 'now' instead of a datetime, since they could use it directly for every purpose 'now' served.